LayoutTests/fast/forms/datalist-expected.txt

Test for the datalist element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS datalist.options.length is 4
PASS datalist.options.length is 3
PASS successfullyParsed is true

TEST COMPLETE
